+++ resources/js/components/admin/pages/AdminReview/AdminReview.jsx
@@ -3,6 +3,7 @@ import { Table } from 'antd';
 import getColumns from './columns';
 import Add from './Add';
 import Edit from './Edit';
+import Swal from 'sweetalert2';
 
 const AdminReview = () => {
 
@@ -15,7 +16,7 @@ const AdminReview = () => {
         await axios.get('/api/reviews')
             .then(({ data }) => {
                 setReviews(data.reviews);
-                console.log(data.reviews);
+
             })
             .catch(error => { throw new Error('Can`t load url') })
     }
@@ -29,17 +30,30 @@ const AdminReview = () => {
     }
 
     const removeReview = async (id) => {
-        setReviews(reviews.filter(p => p.id !== id));
-        const response = await axios.delete('/api/reviews/' + id);
+
+        Swal.fire({
+            title: `Do you want to delete?`,
+            showCancelButton: true,
+            confirmButtonText: 'Yes',
+        }).then((result) => {
+
+            if (result.isConfirmed) {
+                axios.delete('/api/reviews/' + id)
+                    .then(({ data }) => {
+                        setReviews(reviews.filter(p => p.id !== id));
+                        Swal.fire({ icon: 'success', title: 'Success' })
+                    });
+            }
+        })
     }
 
     const editReview = async (id, values) => {
 
         const { data } = await axios.put('/api/reviews/' + id, values);
-        console.log("data in AdminReview: ", data);
+
         const updatedReviews = _.cloneDeep(reviews);
         const review = updatedReviews.find(p => p.id === id);
-        console.log("product in Adminreview: ", review);
+
 
         /*         product.name = data.data.name;
                 product.description = data.data.description;
@@ -49,11 +63,11 @@ const AdminReview = () => {
 
         _.assign(review, data.data)
         setReviews(updatedReviews);
-        console.log("updatedReviews in AdminReview: ", updatedReviews);
+
     }
 
     return (
-        <div className='container'>
+        <div className='admin-review container'>
             <h2 className='my-3'>Reviews</h2>
             <Add
                 addReview={addReview}
